backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
akshare
python-dotenv==1.0.0